HadoopInputFormatIO with junits
